+++ log/include/log/systemsegment.h
@@ -9,7 +9,7 @@ namespace wombat::log {
 
 class SystemSegment : public Segment {
  public:
-  SystemSegment(uint64_t id, const std::filesystem::path& dir, size_t limit);
+  SystemSegment(uint32_t id, const std::filesystem::path& dir, uint32_t limit);
 
   ~SystemSegment() override;
 };
